Update tools.zs
Change stick to forge:rods/wooden to allow treated wood sticks as tool handles (to be consistent with most tools not changed by this script.  Replace recipes for stone and terminite tools to allow stick change and allow any material in the tag minecraft:stone_tool_materials to be used to make stone tools.  Add organization headers to each section.  Change iron and steel plates to use tag to be more compatible with potential future changes.