Update functions.zs
Add changeIngredientsToTag(itemList as IItemStack[], toTag as MCTag<MCItemDefinition>) as void.  Converts all recipes using items in the itemList to recipes using the supplied item tag instead.  Add changeItemListToBaseItem(itemList as IItemStack[], baseItem as IItemStack) as void.  Adds a simple recipe to craft any of the items in the itemList to the baseItem.